Refactor 0 side-effect functions to lib/helpers
