+++ lib/helpers/lerp.js
@@ -1,0 +1,3 @@
+export default function lerp (start, end, ratio){
+    return start + (end-start) * ratio
+}
